requirements.txt
alpaca-py==0.35.0
annotated-types==0.7.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
msgpack==1.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
sseclient-py==1.8.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==14.1
Werkzeug==3.1.3
